add jumphost feature
Introduce a new Jumphost configuration object (as requested in
issue #104) that is available on global level and per device (similar to
username/password). The object contains host/username/password/key_file
to authenticate to device, before connecting through this device to the
end device.

    jumphost:
      host: "jumphost.example.org"
      username: "operator"
      key_file: "jumphost_key"
    devices:
    ...

This patch adds some duplicate code e.g. authForJumphost and could
probably be cleaned up a bit. The code also relies on the keep alive pattern of
the forwarded session rather than managing its own session.

The actual jumphost connection is in `connectToDevice` where the code path is
split in the jumphost and regular connection path.